deleted event dir and moved files to proper places
