+++ corva/app/scheduled.py
@@ -1,7 +1,7 @@
 from corva.app.base import BaseApp
 from corva.models.scheduled import ScheduledContext, ScheduledEventData
 from corva.event import Event
-from corva.event.loader.scheduled import ScheduledLoader
+from corva.loader.scheduled import ScheduledLoader
 from corva.state.redis_adapter import RedisAdapter
 from corva.state.redis_state import RedisState
 from corva.utils import GetStateKey
