+++ corva/app/task.py
@@ -4,7 +4,7 @@ from typing import Literal
 from corva.app.base import BaseApp
 from corva.models.task import TaskData, UpdateTaskInfoData, TaskContext
 from corva.event import Event
-from corva.event.loader.task import TaskLoader
+from corva.loader.task import TaskLoader
 
 
 class TaskApp(BaseApp):
