+++ corva/event.py
@@ -1,15 +1,9 @@
 from __future__ import annotations
 
 from collections import UserList
-from typing import List
-
-from corva.models.base import BaseEventData
 
 
 class Event(UserList):
-    def __init__(self, data: List[BaseEventData]):
-        super().__init__(data)
-
     def __eq__(self, other):
         return (
              super().__eq__(other)
