+++ corva/loader/scheduled.py
@@ -5,7 +5,7 @@ from typing import List
 
 from corva.models.scheduled import ScheduledEventData
 from corva.event import Event
-from corva.event.loader.base import BaseLoader
+from corva.loader.base import BaseLoader
 
 
 class ScheduledLoader(BaseLoader):
@@ -19,4 +19,4 @@ class ScheduledLoader(BaseLoader):
             subdata['app_stream_id'] = subdata.pop('app_stream')
             data.append(ScheduledEventData(**subdata))
 
-        return Event(data=data)
+        return Event(data)
